backend_browse/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
django-cors-headers>=4.3
Pillow>=10.0
